Move shortcut classes to a package
